create new Trajectory file with all equations and removed equations from Lib
